Change list of reg SHM mobj's into tail queue
The OCALL PTA works by, first, requesting the CA allocate shared memory, second,
sending the RPC, and lastly, by requesting the CA free the shared memory. The
shared memory carries the OCALL's memref parameters if any.

The allocation sequence is such that the CA allocates memory, and an entry is
added to reg_shm_list, with guarded = false and the given cookie. Then, the
OCALL PTA resumes and it creates a mobj around that same shared memory with
guarded = true, and the same cookie:

reg_shm_list:
0 (HEAD): guarded = true, cookie = as given
1: guarded = false, cookie = as given

Subsequently, the OCALL PTA sends the RPC for the OCALL proper. Suppose now the
CA crashes when handing the OCALL. Linux cleans up the DMA buf underlying the
shared memory, which prompts a call into OP-TEE to unregister the SHM for the
given cookie. When reg_shm_list is an SLIST, the first element has the right
cookie, but it's the entry that corresponds to the mobj allocated by the OCALL
PTA, not by the CA.

This causes the unregistration to fail, thereby leaking memory in secure world.

Changing the list into a tail queue and by adding new entries at the tail-end
instead of at the head effectively reverses the order of the elements, and fixes
the problem.

Signed-off-by: Hernan Gatta <hegatta@microsoft.com>